fix: `onRemove` callback type (#188)
* fix: onRemove type

* chore: set node16 for ci

* chore: add changeset

* docs: update api

+++ packages/react-native-select-pro/src/components/clear-option/clear-option.hooks.ts
@@ -1,4 +1,5 @@
 import { useSelectContext } from '../../context';
+import { isOptionIndexType, isOptionType } from '../../types';
 
 export const useClearOption = () => {
     const {
@@ -38,7 +39,12 @@ export const useClearOption = () => {
             return;
         }
         removeSingleOption();
-        if (onRemove && selectedOption) {
+        if (
+            onRemove &&
+            selectedOption &&
+            isOptionType(selectedOption) &&
+            isOptionIndexType(selectedOptionIndex)
+        ) {
             // callback
             onRemove(selectedOption, selectedOptionIndex);
         }
